research/requirements.txt
pandas>=2.1
matplotlib>=3.8
plotly>=5.18
geopandas>=0.14
folium>=0.15
shapely>=2.0
pyproj>=3.6
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
sqlalchemy>=2.0
geoalchemy2>=0.14
contextily>=1.7
